remove useAppMode hook file
