+++ src/machines/modelingMachine.ts
@@ -4,7 +4,7 @@ import { isReducedMotion } from 'lang/util'
 import { Axis, Selection, SelectionRangeTypeMap, Selections } from 'useStore'
 import { assign, createMachine } from 'xstate'
 import { v4 as uuidv4 } from 'uuid'
-import { isCursorInSketchCommandRange } from 'hooks/useAppMode'
+import { isCursorInSketchCommandRange } from 'lang/util'
 import {
   doesPipeHaveCallExp,
   getNodePathFromSourceRange,
